- added logging option for better visibility on algo2 workings - fixed settings page skill levels not showing up or updating on change - updated algo2 with various adjustments to work better early on a league

+++ KingsAndQueensHat/Model/Player.cs
@@ -116,7 +116,7 @@ namespace KingsAndQueensHat.Model
 		[XmlIgnore]
 		public decimal WinPercent { get { return GamesPlayed == 0 ? 0 : NumberOfWins/GamesPlayed*100; } }
 		[XmlIgnore]
-		public int AdjustedScore { get; set; }
+		public decimal AdjustedScore { get; set; }
 
         public override string ToString()
         {
